Move elevation computation code into street model
